Fixed slowfalling issue *
Some movement mechanics are still messed up - if you try to jump while moving to the side, the jump quickly gets canceled. Also if you try to move while falling, slow falling activates again. This is because when we move we are setting the rigidbody velocity vector to a new value, therefor entirely cancelling any already existing velocity