changed player movement:
made it so that there was variable jump height, and fixed the problem of being able to jump in the air.
The problem with this is that the gameplay doesn't feel as smooth as before so I'm gonna start working on a double jump to make gameplay feel a little better.