Changed the camera movement - now it's always smooth
CameraFollowCharacter is no longer using lerping to follow the player. Now it moves at the same rate as the player when it's out of the free-moving zone. There's also a range offset parameter now so that you can allow for the range to be shifed up/down or left/right from the player