Death Plane
made it so that there is now a death plane object so that we won't just fall to our death anymore. For now  I just made it a 2d collider, though there might be a better way to do it. Also from now, whenever collision with a certain object is supposed to kill you, I think it'll be most efficient to just call make a new collision event and call the respawn function.